+++ disparity_view/__init__.py
@@ -1,8 +1,8 @@
 from .view import view_npy
 from .view import depth_overlay, as_colorimage
 from .depth_to_normal import DepthToNormalMap
-from .o3d_reprojection import reproject_from_left_and_disparity, gen_right_image
-from .o3d_reprojection import make_animation_gif
+from .o3d_project import reproject_from_left_and_disparity, gen_right_image
+from .o3d_project import make_animation_gif
 
 from .zed_info import (
     get_width_height_fx_fy_cx_cy,
