+++ test/test_o3d_project.py
@@ -15,7 +15,7 @@ import open3d as o3d
 import numpy as np
 import skimage.io
 
-from disparity_view.o3d_reprojection import gen_right_image, make_animation_gif
+from disparity_view.o3d_project import gen_right_image, make_animation_gif
 import inspect
 
 from disparity_view.util import safer_imsave
